microsoft-stack-mastery/microsoft-fabric/requirements.txt
# Microsoft Fabric Python Dependencies
# Install with: pip install -r requirements.txt

# Azure Authentication and Identity
azure-identity>=1.14.0
azure-core>=1.29.0
msal>=1.24.0

# Azure Storage
azure-storage-blob>=12.19.0
azure-storage-file-datalake>=12.13.0

# HTTP and API
requests>=2.31.0

# Configuration Management
python-dotenv>=1.0.0

# Data Processing
pandas>=2.0.0
pyarrow>=14.0.0
numpy>=1.24.0

# PySpark and Delta Lake
pyspark>=3.4.0
delta-spark>=2.4.0

# Database Connectivity
pyodbc>=5.0.0

# Data Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.17.0

# Statistical Analysis
scipy>=1.11.0

# Machine Learning (optional)
scikit-learn>=1.3.0

# Jupyter Notebook Support (optional)
jupyter>=1.0.0
ipykernel>=6.25.0

# Development Tools (optional)
black>=23.9.0
flake8>=6.1.0
pytest>=7.4.0
